fix(project): some of numbers values moved to consts
