+++ src/components/UI/ModalWindows/SubmitProposalModal/SubmitProposalStyles.tsx
