+++ src/utils/consts/numberConsts.ts
@@ -1,0 +1,3 @@
+export const MAX_LENGTH_OF_JOB_DESCRIPTION = 5000;
+export const MIN_LENGTH_OF_JOB_DESCRIPTION = 10;
+export const MAX_LENGTH_OF_COVER_LETTER = 250;
